Extract tests about crypto and key material handling
